refactor: move filterPendingInvitation logic to organizationService
